IA-2488: make generic serializer,viewset for external task
- Make ExternalTaskPostSerializer based on RefreshLqasDataSerializer
- Move object creation code from viewset to serializer
- refactor RefreshLQASDateViewSet
- refactor lqas data refresh serializers: add a GET and a POST serializer
